refactor(grok): isolate the xAI ACP fallback
Keep the shared ACP session runtime provider-neutral and decorate it with Grok's private prompt-completion behavior at the provider boundary. Reject late settlements unless both session and turn lineage match.
