+++ apps/server/src/provider/acp/AcpSessionRuntime.ts
@@ -11,7 +11,6 @@ import * as Queue from "effect/Queue";
 import * as Ref from "effect/Ref";
 import * as Scope from "effect/Scope";
 import * as Semaphore from "effect/Semaphore";
-import * as Schema from "effect/Schema";
 import * as Stream from "effect/Stream";
 import * as ChildProcess from "effect/unstable/process/ChildProcess";
 import * as ChildProcessSpawner from "effect/unstable/process/ChildProcessSpawner";
@@ -40,21 +39,6 @@ function formatConfigOptionValue(value: string | boolean): string {
   return JSON.stringify(value);
 }
 
-const XAiPromptCompleteNotification = Schema.Struct({
-  sessionId: Schema.String,
-  promptId: Schema.optional(Schema.String),
-  stopReason: Schema.optional(Schema.String),
-  agentResult: Schema.optional(Schema.NullOr(Schema.Unknown)),
-});
-
-type XAiPromptCompleteNotification = typeof XAiPromptCompleteNotification.Type;
-
-interface PendingXAiPromptCompletion {
-  readonly sessionId: string;
-  readonly promptId: string;
-  readonly deferred: Deferred.Deferred<EffectAcpSchema.PromptResponse>;
-}
-
 export interface AcpSessionEventStreamBarrier {
   readonly _tag: "EventStreamBarrier";
   readonly acknowledge: Deferred.Deferred<void>;
@@ -62,10 +46,8 @@ export interface AcpSessionEventStreamBarrier {
 
 export type AcpSessionRuntimeEvent = AcpParsedSessionEvent | AcpSessionEventStreamBarrier;
 
-const completedXAiPromptIdLimit = 128;
 const defaultSessionLoadTimeout = Duration.seconds(90);
 const defaultSessionLoadReplayIdleGap = Duration.seconds(2);
-const xAiStopReasonMissingMetaKey = "xAiStopReasonMissing";
 
 export interface AcpSpawnInput {
   readonly command: string;
@@ -303,16 +285,7 @@ export const make = (
     const activePromptFiberRef = yield* Ref.make<
       Option.Option<Fiber.Fiber<EffectAcpSchema.PromptResponse, EffectAcpErrors.AcpError>>
     >(Option.none());
-    const pendingXAiPromptCompletionsRef = yield* Ref.make<
-      ReadonlyArray<PendingXAiPromptCompletion>
-    >([]);
-    const completedXAiPromptIdsRef = yield* Ref.make<ReadonlyArray<string>>([]);
     const sessionLoadGateRef = yield* Ref.make<Option.Option<SessionLoadGate>>(Option.none());
-    let nextXAiPromptFallbackId = 0;
-    const allocateXAiPromptFallbackId = Effect.sync(() => {
-      nextXAiPromptFallbackId += 1;
-      return `t3-xai-prompt-${nextXAiPromptFallbackId}`;
-    });
 
     const logRequest = (event: AcpSessionRequestLogEvent) =>
       options.requestLogger ? options.requestLogger(event) : Effect.void;
@@ -418,17 +391,6 @@ export const make = (
         });
       }),
     );
-    yield* acp.handleExtNotification(
-      "_x.ai/session/prompt_complete",
-      XAiPromptCompleteNotification,
-      (notification) =>
-        resolveXAiPromptCompletionFallback({
-          pendingRef: pendingXAiPromptCompletionsRef,
-          completedPromptIdsRef: completedXAiPromptIdsRef,
-          notification,
-        }),
-    );
-
     const initializeClientCapabilities = {
       fs: {
         readTextFile: false,
@@ -746,60 +708,32 @@ export const make = (
         promptSerializationSemaphore.withPermit(
           Effect.gen(function* () {
             const started = yield* getStartedState;
-            const promptId = yield* allocateXAiPromptFallbackId;
             yield* closeActiveAssistantSegment({
               queue: eventQueue,
               assistantSegmentRef,
             });
-            const fallback = yield* registerXAiPromptCompletionFallback(
-              pendingXAiPromptCompletionsRef,
-              started.sessionId,
-              promptId,
-            );
             const requestPayload = {
               sessionId: started.sessionId,
               ...payload,
-              _meta: {
-                ...payload._meta,
-                promptId: fallback.promptId,
-                requestId: fallback.promptId,
-              },
             } satisfies EffectAcpSchema.PromptRequest;
-            const cancelledResponse = promptResponseFromXAi({
-              sessionId: started.sessionId,
-              promptId: fallback.promptId,
+            const cancelledResponse = {
               stopReason: "cancelled",
-              agentResult: null,
-            });
-            // Grok can emit its private prompt-complete notification even when the
-            // standard session/prompt RPC never settles. Race the RPC with that
-            // notification so callers still emit turn completion and clear UI state.
+            } satisfies EffectAcpSchema.PromptResponse;
             const promptRpcFiber = yield* runLoggedRequest(
               "session/prompt",
               requestPayload,
               acp.agent.prompt(requestPayload),
             ).pipe(Effect.forkIn(runtimeScope));
             yield* Ref.set(activePromptFiberRef, Option.some(promptRpcFiber));
-            return yield* Effect.raceFirst(
-              Fiber.join(promptRpcFiber).pipe(
-                Effect.catchCause((cause) =>
-                  Cause.hasInterruptsOnly(cause)
-                    ? Effect.succeed(cancelledResponse)
-                    : Effect.failCause(cause),
-                ),
-              ),
-              Deferred.await(fallback.deferred),
-            ).pipe(
-              Effect.tap((response) =>
-                rememberCompletedXAiPromptId(completedXAiPromptIdsRef, response, fallback.promptId),
+            return yield* Fiber.join(promptRpcFiber).pipe(
+              Effect.catchCause((cause) =>
+                Cause.hasInterruptsOnly(cause)
+                  ? Effect.succeed(cancelledResponse)
+                  : Effect.failCause(cause),
               ),
               Effect.ensuring(
                 Effect.gen(function* () {
                   yield* Fiber.interrupt(promptRpcFiber).pipe(Effect.ignore);
-                  yield* unregisterXAiPromptCompletionFallback(
-                    pendingXAiPromptCompletionsRef,
-                    fallback.deferred,
-                  );
                   yield* Ref.set(activePromptFiberRef, Option.none());
                 }),
               ),
@@ -819,7 +753,6 @@ export const make = (
             if (Option.isSome(activePromptFiber)) {
               yield* Fiber.interrupt(activePromptFiber.value).pipe(Effect.ignore);
             }
-            yield* abortPendingPromptCompletions(pendingXAiPromptCompletionsRef, started.sessionId);
             yield* acp.agent
               .cancel({ sessionId: started.sessionId })
               .pipe(Effect.ignore, Effect.forkIn(runtimeScope));
@@ -1053,174 +986,3 @@ const closeActiveAssistantSegment = ({
       } satisfies AcpAssistantSegmentState,
     ] as const;
   }).pipe(Effect.flatMap((event) => (event ? Queue.offer(queue, event) : Effect.void)));
-
-const registerXAiPromptCompletionFallback = (
-  pendingRef: Ref.Ref<ReadonlyArray<PendingXAiPromptCompletion>>,
-  sessionId: string,
-  promptId: string,
-) =>
-  Deferred.make<EffectAcpSchema.PromptResponse>().pipe(
-    Effect.tap((deferred) =>
-      Ref.update(pendingRef, (pending) => [...pending, { sessionId, promptId, deferred }]),
-    ),
-    Effect.map((deferred) => ({ deferred, promptId })),
-  );
-
-const unregisterXAiPromptCompletionFallback = (
-  pendingRef: Ref.Ref<ReadonlyArray<PendingXAiPromptCompletion>>,
-  deferred: Deferred.Deferred<EffectAcpSchema.PromptResponse>,
-) => Ref.update(pendingRef, (pending) => pending.filter((entry) => entry.deferred !== deferred));
-
-const abortPendingPromptCompletions = (
-  pendingRef: Ref.Ref<ReadonlyArray<PendingXAiPromptCompletion>>,
-  sessionId: string,
-) =>
-  Ref.modify(pendingRef, (pending) => {
-    const [toAbort, remaining] = pending.reduce<
-      [ReadonlyArray<PendingXAiPromptCompletion>, ReadonlyArray<PendingXAiPromptCompletion>]
-    >(
-      ([aborting, kept], entry) =>
-        entry.sessionId === sessionId ? [[...aborting, entry], kept] : [aborting, [...kept, entry]],
-      [[], []],
-    );
-    if (toAbort.length === 0) {
-      return [Effect.void, pending] as const;
-    }
-    return [
-      Effect.forEach(
-        toAbort,
-        (entry) =>
-          Deferred.succeed(
-            entry.deferred,
-            promptResponseFromXAi({
-              sessionId: entry.sessionId,
-              promptId: entry.promptId,
-              stopReason: "cancelled",
-              agentResult: null,
-            }),
-          ),
-        { concurrency: "unbounded" },
-      ).pipe(Effect.asVoid),
-      remaining,
-    ] as const;
-  }).pipe(Effect.flatten);
-
-const resolveXAiPromptCompletionFallback = ({
-  pendingRef,
-  completedPromptIdsRef,
-  notification,
-}: {
-  readonly pendingRef: Ref.Ref<ReadonlyArray<PendingXAiPromptCompletion>>;
-  readonly completedPromptIdsRef: Ref.Ref<ReadonlyArray<string>>;
-  readonly notification: XAiPromptCompleteNotification;
-}) =>
-  Ref.get(completedPromptIdsRef).pipe(
-    Effect.flatMap((completedPromptIds) => {
-      if (
-        notification.promptId !== undefined &&
-        completedPromptIds.includes(notification.promptId)
-      ) {
-        return Effect.void;
-      }
-      return Ref.modify(pendingRef, (pending) => {
-        const index =
-          notification.promptId !== undefined
-            ? pending.findIndex(
-                (entry) =>
-                  entry.sessionId === notification.sessionId &&
-                  entry.promptId === notification.promptId,
-              )
-            : (() => {
-                const sessionPendingIndexes = pending.flatMap((entry, entryIndex) =>
-                  entry.sessionId === notification.sessionId ? [entryIndex] : [],
-                );
-                if (sessionPendingIndexes.length === 0) {
-                  return -1;
-                }
-                // xAI may omit promptId while multiple steered prompts are in flight.
-                // Resolve the oldest pending fallback FIFO for this session.
-                return sessionPendingIndexes[0] ?? -1;
-              })();
-        if (index < 0) {
-          return [Effect.void, pending] as const;
-        }
-        const entry = pending[index];
-        if (!entry) {
-          return [Effect.void, pending] as const;
-        }
-        return [
-          Deferred.succeed(entry.deferred, promptResponseFromXAi(notification)).pipe(Effect.asVoid),
-          [...pending.slice(0, index), ...pending.slice(index + 1)],
-        ] as const;
-      }).pipe(Effect.flatten);
-    }),
-  );
-
-const rememberCompletedXAiPromptId = (
-  completedPromptIdsRef: Ref.Ref<ReadonlyArray<string>>,
-  response: EffectAcpSchema.PromptResponse,
-  fallbackPromptId: string,
-) => {
-  const promptId = promptIdFromResponse(response) ?? fallbackPromptId;
-  if (promptId.length === 0) {
-    return Effect.void;
-  }
-  return Ref.update(completedPromptIdsRef, (completedPromptIds) => {
-    if (completedPromptIds.includes(promptId)) {
-      return completedPromptIds;
-    }
-    return [...completedPromptIds, promptId].slice(-completedXAiPromptIdLimit);
-  });
-};
-
-function promptIdFromResponse(response: EffectAcpSchema.PromptResponse): string | undefined {
-  const meta = response._meta;
-  if (meta === null || typeof meta !== "object") {
-    return undefined;
-  }
-  const promptId = meta.promptId ?? meta.requestId;
-  return typeof promptId === "string" && promptId.length > 0 ? promptId : undefined;
-}
-
-export function promptResponseHasMissingXAiStopReason(
-  response: EffectAcpSchema.PromptResponse,
-): boolean {
-  const meta = response._meta;
-  return meta !== null && typeof meta === "object" && meta[xAiStopReasonMissingMetaKey] === true;
-}
-
-function promptResponseFromXAi(
-  notification: XAiPromptCompleteNotification,
-): EffectAcpSchema.PromptResponse {
-  const stopReason = normalizeXAiStopReason(notification.stopReason);
-  const meta: Record<string, unknown> = {
-    sessionId: notification.sessionId,
-  };
-  if (notification.stopReason === undefined) {
-    meta[xAiStopReasonMissingMetaKey] = true;
-  }
-  if (notification.promptId !== undefined) {
-    meta.promptId = notification.promptId;
-    meta.requestId = notification.promptId;
-  }
-  if (notification.agentResult !== undefined) {
-    meta.agentResult = notification.agentResult;
-  }
-  return {
-    stopReason,
-    _meta: meta,
-  };
-}
-
-function normalizeXAiStopReason(value: string | undefined): EffectAcpSchema.StopReason {
-  switch (value) {
-    case "cancelled":
-    case "end_turn":
-    case "max_tokens":
-    case "max_turn_requests":
-    case "refusal":
-      return value;
-    default:
-      return "end_turn";
-  }
-}
